pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "a telegram bot for openai chatgpt"
authors = ["sirlaurie <liusuzhang@gmail.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.2.3"
python-telegram-bot = "^20.6"
pymupdf = "^1.23.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
